# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/aluno/registro":{"post":{"summary":"Registrar novo aluno","description":"Registra um novo aluno com seus dados, plano e matrícula.","operationId":"register_aluno_registro_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_register_aluno_registro_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/aluno/checkin":{"post":{"summary":"Check-in do aluno","description":"Registra um check-in do aluno na academia.","operationId":"checkin_aluno_checkin_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckinPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/aluno/checkout":{"post":{"summary":"Checkout","operationId":"checkout_aluno_checkout_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Checkouts"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseStatus"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/aluno/{id}/frequencia":{"get":{"summary":"Relatório com Histórico de Frequência","description":"Este endpoint gera um relatório com o histórico de frequência de um aluno.","operationId":"get_frequency_aluno__id__frequencia_get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","description":"ID do aluno","title":"Id"},"description":"ID do aluno"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/aluno/{id}/risco-churn":{"get":{"summary":"Probabilidade de Churn","description":"Este endpoint retorna a probabilidade de um aluno cancelar sua inscrição.","operationId":"risco_churn_aluno__id__risco_churn_get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","description":"ID do aluno","title":"Id"},"description":"ID do aluno"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Aluno":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"age":{"type":"integer","title":"Age"}},"type":"object","required":["id","name","age"],"title":"Aluno"},"Body_register_aluno_registro_post":{"properties":{"aluno":{"$ref":"#/components/schemas/Aluno"},"plano":{"$ref":"#/components/schemas/Planos"},"matricula":{"$ref":"#/components/schemas/Matriculas"}},"type":"object","required":["aluno","plano","matricula"],"title":"Body_register_aluno_registro_post"},"CheckinPayload":{"properties":{"id_aluno":{"type":"integer","title":"Id Aluno"}},"type":"object","required":["id_aluno"],"title":"CheckinPayload"},"Checkouts":{"properties":{"id":{"type":"integer","title":"Id"}},"type":"object","required":["id"],"title":"Checkouts"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Matriculas":{"properties":{"attender_id":{"type":"integer","title":"Attender Id"},"plan_id":{"type":"integer","title":"Plan Id"},"status":{"type":"boolean","title":"Status"}},"type":"object","required":["attender_id","plan_id","status"],"title":"Matriculas"},"Planos":{"properties":{"name":{"type":"string","title":"Name"},"type":{"type":"string","title":"Type"},"price":{"type":"number","title":"Price"}},"type":"object","required":["name","type","price"],"title":"Planos"},"ResponseStatus":{"properties":{"status":{"type":"string","title":"Status"},"mensagem":{"type":"string","title":"Mensagem"}},"type":"object","required":["status","mensagem"],"title":"ResponseStatus"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}